refactor: split messenger webhook handlers and helpers
